# backend/internal/transport/http/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Racion API",
    "version": "1.0.0",
    "summary": "Weekly meal plans with real store prices and one shopping list.",
    "description": "Read-only catalogue plus plan building. No key, no account: send a request, get JSON.\n\nThe planner is deterministic — the same answers give the same week, and every dish carries the reason it was chosen. Prices are estimates from national statistics (Rosstat in Russia, Eurostat price levels with HICP in Europe, BLS in the US) multiplied by a store index; they are never live shelf prices and must be presented as estimates.\n\nRecipes and photos are ours and licensed for personal, non-commercial use. Building a product on top of this API is fine; republishing the recipe corpus is not.\n\nRate limit per IP address: 600 requests a minute to /api/ overall, of which 60 a minute may build plans (bursts of 100 and 20 are allowed). Over the limit the answer is 429 with a Retry-After header. CORS is open, so the API can be called straight from a browser. Language comes from ?lang=; prices from ?country= and the chosen store.",
    "contact": {
      "name": "Racion",
      "email": "info@racion.app",
      "url": "https://racion.app"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    }
  },
  "servers": [
    {
      "url": "https://racion.app",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/api/plans": {
      "post": {
        "tags": [
          "plans"
        ],
        "operationId": "createPlan",
        "summary": "Build a week of meals with a shopping list",
        "description": "The main call. Give who eats, where they shop and what they cannot have; get seven days of dishes, per-day nutrition, and one shopping list rounded to whole packs with an estimated total.\n\nThe result is stored and readable at GET /api/plans/{id}; the same id opens the human page at /plan/{id}, which is what you show a person.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Params"
              },
              "examples": {
                "family": {
                  "summary": "Two adults and a child, Пятёрочка, no fish",
                  "value": {
                    "country": "RU",
                    "store": "pyaterochka",
                    "region": "",
                    "adults": 2,
                    "kids": [
                      {
                        "age": 7
                      }
                    ],
                    "slots": [
                      "breakfast",
                      "lunch",
                      "dinner"
                    ],
                    "equipment": [
                      "stove",
                      "oven"
                    ],
                    "excludeTags": [
                      "fish"
                    ],
                    "allergens": [],
                    "budgetMode": "perPersonDay",
                    "budgetValue": 400,
                    "lang": "ru"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadInput"
          },
          "429": {
            "$ref": "#/components/responses/TooMany"
          }
        }
      }
    },
    "/api/plans/{id}": {
      "get": {
        "tags": [
          "plans"
        ],
        "operationId": "getPlan",
        "summary": "Read a plan that was built earlier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/lang"
          }
        ],
        "responses": {
          "200": {
            "description": "The plan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "lang": {
        "name": "lang",
        "in": "query",
        "description": "Interface and recipe language, one of 15 codes. ru by default.",
        "schema": {
          "type": "string",
          "enum": [
            "ru",
            "en",
            "de",
            "fr",
            "es",
            "it",
            "pt",
            "pl",
            "uk",
            "tr",
            "nl",
            "cs",
            "kk",
            "zh",
            "ja"
          ]
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "No such object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadInput": {
        "description": "The request cannot be satisfied: unknown store, empty result after filters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooMany": {
        "description": "Rate limit reached, try later",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Params": {
        "type": "object",
        "description": "Answers that define a plan. Only country and store are really needed; the rest have defaults.",
        "required": [
          "country",
          "store"
        ],
        "properties": {
          "country": {
            "type": "string",
            "example": "RU"
          },
          "store": {
            "type": "string",
            "description": "Store code from /api/meta",
            "example": "pyaterochka"
          },
          "region": {
            "type": "string",
            "description": "Region or city for countries with regional prices; empty means national average"
          },
          "lang": {
            "type": "string",
            "example": "ru",
            "description": "Language of the answer. Takes priority over ?lang= and Accept-Language."
          },
          "adults": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "kids": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "age": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 17
                },
                "appetite": {
                  "type": "string",
                  "enum": [
                    "small",
                    "normal",
                    "big"
                  ]
                }
              }
            }
          },
          "slots": {
            "type": "array",
            "description": "Which meals to plan",
            "items": {
              "type": "string",
              "enum": [
                "breakfast",
                "lunch",
                "dinner",
                "snack"
              ]
            }
          },
          "equipment": {
            "type": "array",
            "description": "What the kitchen has: stove, oven, microwave, airfryer, multicooker, blender, mixer, grill, steamer, meatgrinder",
            "items": {
              "type": "string"
            }
          },
          "allergens": {
            "type": "array",
            "description": "Allergen ids from /api/meta; dishes containing them are never offered",
            "items": {
              "type": "string"
            }
          },
          "exclude": {
            "type": "array",
            "description": "Ingredient ids the person will not eat",
            "items": {
              "type": "string"
            }
          },
          "excludeTags": {
            "type": "array",
            "description": "Whole groups: meat, poultry, fish, seafood, offal, spicy",
            "items": {
              "type": "string"
            }
          },
          "goal": {
            "type": "string",
            "description": "none | healthy | lose | gain"
          },
          "budgetMode": {
            "type": "string",
            "enum": [
              "perPersonDay",
              "week"
            ]
          },
          "budgetValue": {
            "type": "number",
            "description": "Target in the currency of the country; 0 means no target"
          },
          "collection": {
            "type": "string",
            "description": "Collection slug: its dishes are taken first"
          },
          "have": {
            "type": "array",
            "description": "Ingredient ids already at home; the shopping list subtracts them",
            "items": {
              "type": "string"
            }
          },
          "startDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "Plan": {
        "type": "object",
        "description": "A built week or table. The same id opens the page for people at /plan/{id}.",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lang": {
            "type": "string"
          },
          "country": {
            "type": "object"
          },
          "store": {
            "type": "object"
          },
          "portions": {
            "type": "number"
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Day"
            }
          },
          "shopping": {
            "type": "array",
            "description": "Shopping list grouped by aisle category",
            "items": {
              "$ref": "#/components/schemas/ShopGroup"
            }
          },
          "totals": {
            "type": "object",
            "properties": {
              "cost": {
                "type": "number",
                "description": "Estimated cost of everything on the list"
              },
              "usedCost": {
                "type": "number",
                "description": "What is actually eaten during the plan; the rest is leftover packs"
              },
              "items": {
                "type": "integer"
              },
              "kcalPerDay": {
                "type": "number"
              },
              "proteinPerDay": {
                "type": "number"
              },
              "cookMin": {
                "type": "integer"
              }
            }
          },
          "priceSource": {
            "$ref": "#/components/schemas/PriceSource"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bots": {
            "type": "object",
            "description": "Links that connect this plan to the Racion bot in a messenger, keyed by messenger (telegram, max). The bot sends the shopping list by aisle; items ticked in the chat are ticked on the plan page too. Only messengers the server runs a bot for are listed.",
            "additionalProperties": {
              "type": "string",
              "format": "uri"
            }
          }
        }
      },
      "Day": {
        "type": "object",
        "properties": {
          "index": {
            "type": "integer"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "label": {
            "type": "string"
          },
          "kcal": {
            "type": "number"
          },
          "cost": {
            "type": "number"
          },
          "dishes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dish"
            }
          }
        }
      },
      "Dish": {
        "type": "object",
        "properties": {
          "slot": {
            "type": "string"
          },
          "recipeId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "timeMin": {
            "type": "integer"
          },
          "kcal": {
            "type": "number"
          },
          "cost": {
            "type": "number",
            "description": "Per portion"
          },
          "leftover": {
            "type": "boolean",
            "description": "Yesterday's dish eaten again"
          },
          "batch": {
            "type": "boolean",
            "description": "Cooked once for two days"
          },
          "servings": {
            "type": "integer",
            "description": "Baskets only: how many people this dish feeds"
          },
          "why": {
            "type": "string",
            "description": "Why the planner chose it: finishing an open pack, under budget, fits calories"
          }
        }
      },
      "ShopGroup": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "description": "Aisle code: vegetables, meat, dairy…"
          },
          "label": {
            "type": "string",
            "description": "Aisle name in the plan language"
          },
          "cost": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ingredientId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "needed": {
                  "type": "number",
                  "description": "How much the recipes actually use"
                },
                "buy": {
                  "type": "number",
                  "description": "How much to buy: needed rounded up to whole packs"
                },
                "pack": {
                  "type": "number",
                  "description": "Pack size sold in shops"
                },
                "unit": {
                  "type": "string",
                  "enum": [
                    "g",
                    "ml",
                    "pcs"
                  ]
                },
                "cost": {
                  "type": "number"
                },
                "pantry": {
                  "type": "boolean",
                  "description": "Usually already at home; not counted in the total"
                },
                "usedIn": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "Dishes this goes into"
                }
              }
            }
          }
        }
      },
      "PriceSource": {
        "type": "object",
        "description": "Where the price estimate came from. Show it whenever you quote a figure.",
        "properties": {
          "name": {
            "type": "string",
            "example": "Eurostat: country price level × HICP by group"
          },
          "period": {
            "type": "string",
            "example": "August 2026"
          },
          "region": {
            "type": "string"
          },
          "coverage": {
            "type": "number",
            "description": "Share of the basket priced from the source, 0..1"
          }
        }
      }
    }
  }
}